+++ R/display.R
@@ -62,6 +62,13 @@
 #'   select = "{coef}{stars}|({ci})",
 #'   column_labels = c("Estimate", "95% CI")
 #' )
+#'
+#' # line break, unicode minus-sign
+#' print_html(
+#'   out,
+#'   select = "{estimate}{stars}<br>({ci_low} \u2212 {ci_high})",
+#'   column_labels = c("Est. (95% CI)")
+#' )
 #' }
 #' @export
 display.parameters_model <- function(object,
@@ -107,9 +114,6 @@ display.parameters_model <- function(object,
     )
   }
 }
-
-#' @export
-display.parameters_stan <- display.parameters_model
 
 #' @export
 display.parameters_simulate <- display.parameters_model