+++ R/methods_base.R
@@ -136,10 +136,10 @@ p_value.numeric <- function(model, null = 0, ...) {
 
 #' @export
 p_value.data.frame <- function(model, ...) {
-  data <- model[sapply(model, is.numeric)]
+  data <- model[vapply(model, is.numeric, TRUE)]
   .data_frame(
     Parameter = names(data),
-    p = sapply(data, p_value)
+    p = vapply(data, p_value, 1)
   )
 }