+++ R/standardize_info.R
@@ -50,7 +50,7 @@ standardize_info.default <- function(model,
   types <- parameters_type(model)
   # model_matrix <- as.data.frame(stats::model.matrix(model))
   model_matrix <- as.data.frame(insight::get_modelmatrix(model))
-  data <- insight::get_data(model)
+  data <- insight::get_data(model, source = "mf", verbose = FALSE)
   wgts <- insight::get_weights(model, na_rm = TRUE)
 
   # Sanity Check for ZI
@@ -76,8 +76,8 @@ standardize_info.default <- function(model,
 
   # Type of effect size
   out$EffectSize_Type <- ifelse(types$Type == "interaction", "interaction",
-    ifelse(types$Link == "Association", "r",
-      ifelse(types$Link == "Difference", "d", NA)
+    ifelse(types$Link == "Association", "r", # nolint
+      ifelse(types$Link == "Difference", "d", NA) # nolint
     )
   )
 
@@ -118,9 +118,7 @@ standardize_info.default <- function(model,
   )
 
   # Pseudo (for LMM)
-  if (include_pseudo &&
-    mi$is_mixed &&
-    length(insight::find_random(model)$random) == 1) {
+  if (include_pseudo && mi$is_mixed && length(insight::find_random(model)$random) == 1L) {
     out <- merge(
       out,
       .std_info_pseudo(
@@ -141,7 +139,7 @@ standardize_info.default <- function(model,
   row.names(out) <- NULL
 
   # Remove all means for now (because it's not used)
-  out <- out[!grepl("Mean_", names(out))]
+  out <- out[!grepl("Mean_", names(out), fixed = TRUE)]
 
   # Select only desired columns
   # if(method == "all") method <- c("smart", "basic")
@@ -436,7 +434,7 @@ standardize_info.default <- function(model,
     if (length(also_between)) {
       insight::format_warning(
         "The following within-group terms have between-group variance:",
-        paste0(also_between, collapse = ", "),
+        toString(also_between),
         "This can inflate standardized within-group parameters associated with these terms.",
         "See `help(\"demean\", package = \"datawizard\")` for modeling between- and within-subject effects."
       )