+++ tests/testthat/test-model_parameters.cgam.R
@@ -1,6 +1,6 @@
 .runThisTest <- Sys.getenv("RunAllparametersTests") == "yes"
 
-if (.runThisTest && requiet("testthat") && requiet("cgam") &&
+if (.runThisTest && requiet("cgam") &&
   utils::packageVersion("insight") > "0.18.0" &&
   getRversion() >= "4.0.0") {
   test_that("model_parameters - cgam", {
@@ -137,6 +137,10 @@ if (.runThisTest && requiet("testthat") && requiet("cgam") &&
         ran_pars = TRUE,
         show_summary = FALSE,
         log_link = FALSE,
+        pretty_labels = c(
+          `(Intercept)` = "(Intercept)",
+          `s.incr(x)` = "s.incr(x)"
+        ),
         test_statistic = "t-statistic",
         log_response = FALSE,
         log_predictors = FALSE,